ENH: Use a common, backend-independent implementation of ColorColumn.
